Revert "Merge branch 'feature/physics' into develop"
This reverts commit e7cc7b452cfdb5a70da5333a9f99b1c245897b58.

+++ GameEninge/Transform.hpp
@@ -10,8 +10,8 @@ namespace spic {
      */
     struct Transform {
         Point position; // Translation (shift)
-        float rotation; // Rotation, in radians
-        float scale; // Multiplication factor
+        double rotation; // Rotation, in radians
+        double scale; // Multiplication factor
     };
 }
